+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.127.0] - 2026-07-19
+
 ### Changed
 
 - `gk pr` / `gk issue` / `gk inbox` are **interactive by default in a terminal**,